requirements.txt
Flask==2.0.2
Flask-RQ2==18.3
flask-sse==1.0.0
redis==3.5.3
redis==3.5.3
pytube==11.0.1
google-cloud-storage==1.29.0
gunicorn==20.1.0
gevent==21.8.0
assemblyai==0.2.7